+++ src/shutoushu/model/shutoushu-routing-protocol.cc
@@ -991,8 +991,9 @@ RoutingProtocol::RecvShutoushu (Ptr<Socket> socket)
             if (m_finish_time[des_id] == 0)
               {
                 m_finish_time[des_id] = Simulator::Now ().GetMicroSeconds ();
-                packetTrajectory << send_x << ", " << send_y << ", " << mypos.x << ", " << mypos.y
-                                 << ", " << Simulator::Now ().GetMicroSeconds () << ", "
+                packetTrajectory << send_x << ", " << send_y << ", " << (int) mypos.x << ", "
+                                 << (int) mypos.y << ", " << Simulator::Now ().GetMicroSeconds ()
+                                 << ", "
                                  << "destination"
                                  << ", " << hopcount << ", " << id << ", " << send_id << ", "
                                  << des_id << ", " << std::endl;
@@ -1040,10 +1041,11 @@ RoutingProtocol::RecvShutoushu (Ptr<Socket> socket)
                     // //std::cout << "\n--------------------------------------------------------\n";
                     std::cout << "関係あるrecv id" << id << "time------------------------------\n"
                               << Simulator::Now ().GetMicroSeconds ();
-                    packetTrajectory << send_x << ", " << send_y << ", " << mypos.x << ", "
-                                     << mypos.y << ", " << Simulator::Now ().GetMicroSeconds ()
-                                     << ", " << i << ", " << hopcount << ", " << id << ", "
-                                     << send_id << ", " << des_id << ", " << std::endl;
+                    packetTrajectory << send_x << ", " << send_y << ", " << (int) mypos.x << ", "
+                                     << (int) mypos.y << ", "
+                                     << Simulator::Now ().GetMicroSeconds () << ", " << i << ", "
+                                     << hopcount << ", " << id << ", " << send_id << ", " << des_id
+                                     << ", " << std::endl;
                     SendShutoushuBroadcast (i + 1, des_id, des_x, des_y, hopcount);
                   }
                 else //含まれていないか
@@ -1061,10 +1063,11 @@ RoutingProtocol::RecvShutoushu (Ptr<Socket> socket)
                     //std::cout << "\n--------------------------------------------------------\n";
                     std::cout << "関係あるrecv id" << id << "time------------------------------"
                               << Simulator::Now ().GetMicroSeconds () << "\n";
-                    packetTrajectory << send_x << ", " << send_y << ", " << mypos.x << ", "
-                                     << mypos.y << ", " << Simulator::Now ().GetMicroSeconds ()
-                                     << ", " << i << ", " << hopcount << ", " << id << ", "
-                                     << send_id << ", " << des_id << ", " << std::endl;
+                    packetTrajectory << send_x << ", " << send_y << ", " << (int) mypos.x << ", "
+                                     << (int) mypos.y << ", "
+                                     << Simulator::Now ().GetMicroSeconds () << ", " << i << ", "
+                                     << hopcount << ", " << id << ", " << send_id << ", " << des_id
+                                     << ", " << std::endl;
                     SendShutoushuBroadcast (i + 1, des_id, des_x, des_y, hopcount);
                   }
                 else //含まれていないか
